Panel displays picture

+++ proj/src/perlinNoise/NoiseInterpreter.java
@@ -51,9 +51,9 @@ public class NoiseInterpreter {
 	private static Color getGradientColor(Color gradientStart, Color gradientEnd, float value) {
 		float inverse = 1 - value;
 		Color c = new Color(
-				gradientStart.getRed() * inverse + gradientEnd.getRed() * value,
-				gradientStart.getGreen() * inverse + gradientEnd.getGreen() * value,
-				gradientStart.getBlue() * inverse + gradientEnd.getBlue() * value);		
+				(int)(gradientStart.getRed() * inverse + gradientEnd.getRed() * value),
+				(int)(gradientStart.getGreen() * inverse + gradientEnd.getGreen() * value),
+				(int)(gradientStart.getBlue() * inverse + gradientEnd.getBlue() * value));		
 		return c;
 	}
